ril: avoid sending null aid string in getIMSI
A change to how getIMSI works in JB causes null aid strings
to be sent.  The result for RILs older than v7 is a corrupt
ril response in SIMRecords::fetchSimRecords.  This corruption
ends up setting mImsi to null (in SIMRecords) and that causes
havoc w/ 3g/4g data connections.

This new patch is implemented as an opt-in old ril feature

Change-Id: I6682ea10c8b216e2adc5abffbcc50f3cb6c9d7b8