frameworks/base:(Hack)Print the content displayed with Canvas.drawText
This hack allows us to print content that displayed with the drawText
method of Canvas.java when ro.debug.drawtext is set to true.

This allows us to capture the contents on the screeen
which will help us capture benchmark results for thirdparty apps
which use drawText of Canvas to display results.

Change-Id: Ic114fbe8d469ff27df57c585d3af322c703b4b83
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>