TORCH FIX
Made torch check for a mismatch between torch status and checked-box status1 second after the toggle changes, in addition to checking when a sharedpref change happens (in case it doesn't).

Now, doesn't get stuck in a disabled state, even when screen locked, and in the worst case, takes 2 toggles to affect change in torch status.

Change-Id: I987e77e195da07ec7f026dd2fddaea6dec028ee6
Signed-off-by: nuclearmistake <nuclearmistake@gmail.com>