Default device volumes not initialised
After fresh rom install default device volumes are not initialised
and therefore muted.

Only after manually changing a device volume it can be used.
IMHO the check for AudioSystem.DEVICE_OUT_DEFAULT here is not correct
since it will not init mIndex and mLastAudibleIndex collections.

Maybe there is a better way to fix this

Change-Id: I00ab1a100528fa624dae0498d8c8a41a35a59a78