Fix battery charging image themechooser compatibility.
It's a shame, because I liked this, but changing the names of and adding these drawables makes themechooser drawables not replace ours properly. Themechooser replaces drawables by matching file names... even if we just added png's between the stock ones, we'd at best be alternating between those defined by a themechooser theme and our added ones.

On crystal2 theme, for example, when the charging animation is nearly full, it jumps back down to about 10%, then shoots back up to finish filling... this does not occur on the "system" theme, which uses our drawables exclusively.

Squashed commit of the following:

commit 430e79666a9eddbccd844bcf34ed64b49f53559e
Author: nuclearmistake <nuclearmistake@gmail.com>
Date:   Thu Jan 17 08:54:01 2013 -0500

    Revert "Added more battery charging pngs to make the animation smoother"

    This reverts commit 36aa5de0be581e5290d65e2e10984a2dc726212c.

commit ffb8156440748c1b3b2dc1d73a6fe4dd9d05e2bb
Author: nuclearmistake <nuclearmistake@gmail.com>
Date:   Thu Jan 17 08:53:53 2013 -0500

    Revert "Fine tuned the drawables. Nothing major."

    This reverts commit 252aa88bdf6bee820a3867eb9a27a0ce81ac45ec.

Change-Id: I0fde53964ae45a4edcff474620a28974c0614c8a