navbar: remove hashmaps in favor of static enum constants
this is a pretty heavily modified version of the AOKP navbar. the features
we want to include are all present and improved on accordingly.

PS1: refactor. add last app (broken?) removed quicksetting dependencies on
vanirawesome giving it it's own action class

PS2: made sure systemUI not get killed again and removed a dead screenshot derp

PS3: re-activated screenshot as a broadcast intent and moved it to vanir actions.
     now can be used as a desktop icon and possibly as a quicksetting tile.
     cleaned up QuickAwesome. added screenshot images to all resolutions.

PS4: rebase on jb42

Change-Id: Ie8959359790ab3688ce5311bc81874a5c345eb76