QuickSettings : Battery initial status
Battery tile upon being added relies on a battery level changed call
to update its status. Thus if upon adding, the battery level is not changing
or till the period when it updates, the tile shows an unknown state.

This patchset resolves this issue by making the controller store its last received value
and distributing it to the tiles when they register for changes.

https://jira.cyanogenmod.org/browse/CYAN-1374

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/policy/BatteryController.java
	packages/SystemUI/src/com/android/systemui/statusbar/policy/DockBatteryController.java

Change-Id: I22c353ff49092cda6a0b87a8c4ffcc07a829f33a
Signed-off-by: DHOMD <DHO@dho.im>