Allow ADB to be enabled by default via a config overlay
This is useful for developers that are testing and troubleshooting boot issues
and adb ends up being disabled by default once the system kicks in.  This should
only be used for testing and removed once a device is deemed stable and able to
boot.

PS2:fix wrong default set by default

Change-Id: Ib17d4e10ec68640ce55a91d5b1067d3d37eeec9d