frameworks/base:(Hack)Print the content displayed with Canvas.drawText
This hack allows us to print content that displayed with the drawText
method of Canvas.java when ro.debug.drawtext is set to true.

This allows us to capture the contents on the screeen
which will help us capture benchmark results for thirdparty apps
which use drawText of Canvas to display results.

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>