Active-D: save some B
Only enable the proximity sensor for long enough to get one reading at <1 Hz.

ps3: fix nomenclature error
ps5: don't actually register and unregister the listener... a change that happened while unregistered is not reliably reflected in the values in the callback.
ps90: Change handler the settingsobserver is instantiated with, and don't catch securityexceptions to allow systemui to be resurrected if the first notification-showing fails

Change-Id: I4627cb4ca27b07fac8f1606c02d886a3efa2af40