Scroll: modified velocitytracker
Velocitytracker and overscroll are called on every touch event to
calculate if there is a scrolling/fling event and when the event stops.
Move this and the recycling to static voids so they are no longer processed
after evoking the method call during runtime (JIT) to increase
responsiveness.

PS2: better typedarray handling
PS3: remove throwexception code from add motionevent. throwexception
     is bulky because it has to generate stack trace info and shouldn't
     be necessary here.  if addMovement is called when a motion event is null
     then someone is a complete fuckass and needs to rewrite their shit anyways.

-Vanir

Change-Id: I8c78409f41b8045d41acbeb3d710a324c5267901