update how power reboot menu handles settings changes
Instead of 98,237 getInts on the UI thread as the dialog creates lets
use a broadcast receiver to update all the setting values at once
as any are changed. This also accounts for Stock mode by bypassing the
getInts for non-AOSP settings and using a table of presets instead.

PS2: move airplaneModeOn to OnAirplaneModeChanged() {

-VanirAOSP

Change-Id: I02c69792e7f526e93ca8d15b156fe78bf7c0ac40