ExternalStorageProvider: use mount point in device list
For me its is extremely unlogical that only the volume label is used
in the device list. Even more because it can be an empty string
and for a normal android user this information is not visible
anywhere else.

Using the mount point and if available the device label
separated by "-" makes much more sense and shows the same
information as all other apps that access external storage

Change-Id: I3fc3675af6640e37aa999b03ae2de2858237c638
Signed-off-by: DHOMD <DHO@dho.im>