STE Multimedia Support (4/4)
Fixes "OK Google" on STE devices

-It seems thats the "source" device is defined incorrectly
for this device. So we just switch to the known mic source.

-This might be better fixed in somewhere else but this works
until/if a better fix is found.

-When google now launcher tries to open an mic audio stream
it uses the device id 1999 and expects that to be a mic device.
We were getting an error because 1999 is not either not defined
at all or not defined as a mic.
1 is the device id for our mic so, just switch when it sees 1999.
Change-Id: I8d60b836da988e8313d3d54624d728a2753baf50