Share hardware capability info between QuickSettingsController instances
If these variables are going to be in the heap rather than on the stack (which, given the frequency with which they're used, wouldn't be the end of the world),
we may as well only do it once.

This will not have the problem introduced in http://vaniraosp.goo.im/#/c/9195/ and fixed in aa865ed because all of the values AND the boolean that keeps them from being re-retrieved and re-stored are static.

Change-Id: I0e83cd618822ce42df2c5b97d485c19e90ff2d77