On the go mode squashed
originally from Nameless ROM Author: Evisceration <eviscerationls@gmail.com>
ported to be compatible with vanir with minor improvements

On-The-Go turns on the camera of your phone and streams it to an overlay, which
is drawn over everything. this makes you feel like your display is transparent,
as you see the content behind it.

squash of additional improvement commits:
 http://gerrit.nameless-rom.org/#/c/3593/
 http://gerrit.nameless-rom.org/#/c/3594/
 http://gerrit.nameless-rom.org/#/c/3595/
 http://gerrit.nameless-rom.org/#/c/3724/
 http://gerrit.nameless-rom.org/#/c/3733/5

PS2:Remove extra animation junk that got in during the cherry-pick
PS3:Remove more extra animation junk

Change-Id: Ic0c2fdff0bd7789f55135c288452874e74cc5de3