PhoneStatusBar: remove race condition in adding the navbar view
We are observing the same results in 2 content observers in remote classes here.. so
lets make sure the aidl interface gets the right values from PhoneWindowsManager.
btw, why are we using the aidl interface to gather a boolean value from a remote
class if we're watching the constant in both classes?

Change-Id: I29f58554dda0dca717adcda372a237651fbf28d0