Navbar: alternate layouts
allow up to 5 customizable alternate layouts and add special arrow keys
on each side to switch layouts.  Also introduces a number of optimizations
for setting up the navbar.

to do list:
perhaps use a single string with delimiters per string.. and allow an indefinite
number of layout additions

PS2: make sure layout number is reset if the user is on a higher value than the
amount available
PS3: if the layout number is higher than the new amount of layouts selected set the bar
to the number of the new amount instead of a cleared arraylist

Change-Id: I2bfddb5967a77c63e1d403d1defa01bb3d123d42