fix the race condition in PSB for adding the navbar to hardware devices
Previously the windowsmanagerservice binder call to PhoneWindowsManager was returning a value
set in a massive synchronized block of code that was run after its content observer reported a change.
Since PSB had the same uri registered in its content observer there was a race condition depending
on which side updated first.

Change-Id: Iad9e317b5d72b60deab1f2638320d8bfcec6a1a0