AwesomeConstants: reorder/organize
this needed to be organized and the strings needed to be cleaned up pretty badly.
Torch was moved out of vanir actions and it will now remove itself if the device
doesn't support it (it couldn't remove itself as a vanir action).  Power was moved
to vanir actions in it's place.

Change-Id: Ib0a628a303d61b8a4f3dd946f092ce1c07b0a9c7