AwesomeConstants: rename + renamespace + refactor + update copyright block
AwesomeConstants might as well have Vanir tattooed on its face... such a large tattoo that its family doesn't recognize it at christmas dinner.

De-static some static imports that weren't being used thusly. (though many of the lines in the files with those static imports are on top of a short list.... FOR PAIN)

Change-Id: If6df1ea9842a1880b34c27786e9ef9d444b25fa3