revert forcepushed commit with deleted github history that didn't go through gerrit
i've returned the animator set to being created a minimum of 2 times per onTouchEvent.
Though, i think allocating a trivial ammount of space on the heap for it one time is
better than creating it and making the VM have to push it off the stack over and
over and over.

Change-Id: I1f4203339357cf9a5794c5b704d0e05a33251098