Statusbar: clean up expanded status bar header layout
For future reference: view.invisibile keeps the spacing for layout purposes and
was pushing the "flip to" buttons off the screen with long dates and am/pm enabled.
Also, make the padding between header items consistent and resize button spacing
slightly in case 3 are enabled which also pushes the last button off the screen.

Change-Id: I88cc1302bfe8f32bb7396680704118865c872163