NBV: inline trivial method
Revert "TF this is keeps icons from being the wrong size when the screen rotates. Revert "wtf is this?""

We aren't going to be clearing all the button arrays and redoing them every single layout change.  It's grossly
slower and not necessary. This completes the previous changeset that was reverted.

Change-Id: I27249ee00e60061f467f015d53fc3487d2ef07bd