Revert "Revert "NBV: inline trivial method" and the revert revert that RE-BROKE something."
look at line change 694 in NBV of the string array -> constructor commit.  this commit got split
between the two accidentally.  as it is now it shouldn't break anything.  if it still does, you need
to find a real fix because navbar height is in no way related to the order of the soft keys and we're
not going to reload ALL the button arrays every single layout change.  Doing it that way  is so much
slower it's noticeable on my N5 and it completely defeats the purpose of having them stored in the
first place.

This reverts commit 1bc6cf5338e26dc7df1ae3632d83c68731ccd161.

Change-Id: Ia1a669e3909df94f9f50a8eaad9683ebdeb59ae2