Wifi notifications: notify when wifi connects [1/2]
Options:
  off (default)
  a dry white toast
  a notification
  a notification with a sound

For those of us that survive on wifi islands in small college towns this is
a nice addition

The first notification will be canceled if it happens within 30 seconds of
ACTION_BOOT_COMPLETED to help prevent startup notification spamming. 8 of the
previous connections are stored and it will update the content description
accordingly if you've reconnected to one.  This number can be changed in a
device overlay

Change-Id: I50f00aabad692d2547f4c74bdaaae84947ba84eb