Squash SlimSeekBarPreference's On-The-Go mode 1/2 and introduce VanirActions and Utills for bringing in future features
Frameworks: SlimSeekBarPreference

Patchset: materialize slider preference

Patchset: add licensing info to SlimSeekBarPreference.java

Squashed On-The-Go mode (1/2)

On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

have fun!

Change-Id: Idbf77fc0be2bcd2e9542fd8b6263b362f8a83af4
Signed-off-by: Varun <date.varun123@gmail.com>
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>
Signed-off-by: Ethan Johnson <ej.johnson1999@gmail.com>